Move into_eth_sig back up the dep tree
