+++ node/src/kdf.rs
@@ -1,6 +1,11 @@
-use crypto_shared::ScalarExt;
+use anyhow::Context;
+use crypto_shared::{x_coordinate, ScalarExt, SignatureResponse};
 use hkdf::Hkdf;
-use k256::Scalar;
+use k256::{
+    ecdsa::{RecoveryId, VerifyingKey},
+    elliptic_curve::sec1::ToEncodedPoint,
+    Scalar,
+};
 use near_primitives::hash::CryptoHash;
 use sha2::Sha256;
 
@@ -18,3 +23,38 @@ pub fn derive_delta(receipt_id: CryptoHash, entropy: [u8; 32]) -> Scalar {
 // Constant prefix that ensures delta derivation values are used specifically for
 // near-mpc-recovery with key derivation protocol vX.Y.Z.
 const DELTA_DERIVATION_PREFIX: &str = "near-mpc-recovery v0.1.0 delta derivation:";
+
+// try to get the correct recovery id for this signature by brute force.
+pub fn into_eth_sig(
+    public_key: &k256::AffinePoint,
+    big_r: &k256::AffinePoint,
+    s: &k256::Scalar,
+    msg_hash: Scalar,
+) -> anyhow::Result<SignatureResponse> {
+    let public_key = public_key.to_encoded_point(false);
+    let signature = k256::ecdsa::Signature::from_scalars(x_coordinate(big_r), s)
+        .context("cannot create signature from cait_sith signature")?;
+    let pk0 = VerifyingKey::recover_from_prehash(
+        &msg_hash.to_bytes(),
+        &signature,
+        RecoveryId::try_from(0).context("cannot create recovery_id=0")?,
+    )
+    .context("unable to use 0 as recovery_id to recover public key")?
+    .to_encoded_point(false);
+    if public_key == pk0 {
+        return Ok(SignatureResponse::new(*big_r, *s, 0));
+    }
+
+    let pk1 = VerifyingKey::recover_from_prehash(
+        &msg_hash.to_bytes(),
+        &signature,
+        RecoveryId::try_from(1).context("cannot create recovery_id=1")?,
+    )
+    .context("unable to use 1 as recovery_id to recover public key")?
+    .to_encoded_point(false);
+    if public_key == pk1 {
+        return Ok(SignatureResponse::new(*big_r, *s, 1));
+    }
+
+    anyhow::bail!("cannot use either recovery id (0 or 1) to recover pubic key")
+}
